Modified db-prefill to use `fork()` to start up an instance of the API and persona-faker servers, then kill them on completion; this means that we can now run db-prefill without any prior setup (after `npm install`).
This makes use of redis pub/sub to detect when the API and persona-faker servers are set up, so it requires each to send a message after launching their HTTP server so that db-prefill can begin the pre-fill process. If you have already run `npm install` with persona-faker, you will need to run `npm install` again to receive this change on persona-faker.